requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
streamlit
pymysql
sqlalchemy
chromadb
sentence-transformers
